requirements.txt
click==8.0.4
Flask==2.0.3
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.0
paho-mqtt==1.6.1
Werkzeug==2.0.3
